requirements.txt
pysrt==1.1.2
deep-translator==1.11.4
tqdm==4.65.0
requests==2.31.0
python-dotenv==1.0.0 